chore:Organize Docker container ports alphabetically and import Auth-Guard as a module.Add rimraf as a dev dependency.
