+++ apps/tedrisat/src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { LoggerModule, AuthGuardModule } from '@madrasah/common';
+import { LoggerModule , AuthGuardModule } from '@madrasah/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ConfigModule } from '@nestjs/config';
