+++ apps/tedrisat/src/config/config.ts
@@ -6,7 +6,7 @@ export default () => ({
   serviceName: process.env.SERVICE_NAME || pkg.name,
   version,
   environment: process.env.NODE_ENV || 'development',
-  port: process.env.PORT || 3002,
+  port: process.env.PORT || 3001,
   database: {
     host: process.env.DB_HOST || 'localhost',
     port: process.env.DB_PORT || 5432,
