+++ docker-compose.yml
@@ -7,7 +7,7 @@ services:
       context: .
       dockerfile: apps/tedrisat/Dockerfile
     ports:
-      - "${TEDRISAT_PORT:-3002}:3002"
+      - "${TEDRISAT_PORT:-3001}:3001"
     env_file:
       - ./apps/tedrisat/.env
 
@@ -16,7 +16,7 @@ services:
       context: .
       dockerfile: apps/teskilat/Dockerfile
     ports:
-      - "${TESKILAT_PORT:-3001}:3001"
+      - "${TESKILAT_PORT:-3002}:3002"
     env_file:
       - ./apps/teskilat/.env
