+++ libs/common/src/auth-guard/auth-guard.tokens.ts
@@ -1,0 +1,2 @@
+export const PUBLIC_KEY_PROVIDER = 'IPublicKeyProviderToken';
+export const JWT_VERIFIER = 'IJwtVerifierToken';
