+++ libs/common/src/auth-guard/index.ts
@@ -1,4 +1,2 @@
-export * from './services/jwt-verifier.service';
-export * from './key-providers/dummy-provider';
-export * from './auth-guard.guard';
+export * from './auth-guard';
 export * from './auth-guard.module';
