+++ package.json
@@ -31,7 +31,6 @@
   "license": "ISC",
   "packageManager": "npm@10.2.4",
   "devDependencies": {
-    "@types/jsonwebtoken": "^9.0.10",
     "rimraf": "^6.0.1",
     "turbo": "^2.5.5"
   }
